Feature/api to futures

See merge request !42

+++ app/controllers/FlowInstanceApi.scala
@@ -7,7 +7,11 @@ import controllers.util._
 import global.AuthorisationService
 import org.dcs.api.service.FlowInstance
 import org.dcs.flow.FlowApi
-import play.api.mvc.EssentialAction
+import play.api.mvc.{Action, EssentialAction}
+import global.ResultSerialiserImplicits._
+import play.api.libs.concurrent.Execution.Implicits.defaultContext
+
+import scala.concurrent.Future
 
 /**
   * Created by cmathew on 08/06/16.
@@ -30,54 +34,61 @@ class FlowInstanceApi @Inject()(csrfCheckAction: CSRFCheckAction,
   val ResourceNameSeparator = ":"
   val Type = "http://alambeek.org/resource/instance"
 
-  override def list: EssentialAction =  (csrfCheckAction
-    andThen authorisationAction) { implicit request =>
+  override def list: EssentialAction =  csrfCheckAction andThen authorisationAction async { implicit request =>
     val ids = flowInstanceIds(authService.permissions(request.claims))
     if(ids.isEmpty)
-      serialize(Nil)
+      Future.successful(Nil.toResult)
     else
-      serialize(ids.map( id => FlowApi.instance(id, DefaultUserId, Req.tokenOrError(Req.AuthTokenKey))))
+      Future.sequence(ids.map { id =>
+        FlowApi.instance(id, DefaultUserId, Req.tokenOrError(Req.AuthTokenKey))
+          .map { flowInstance =>
+            flowInstance
+          }
+      }).map(fiList => fiList.toResult)
   }
 
   override def update(id: String): EssentialAction = csrfCheckAction { implicit request =>
     NotImplemented
   }
 
-  override def destroy(id: String): EssentialAction = (csrfCheckAction
-    andThen authorisationAction) { implicit request =>
-    val result = FlowApi.remove(id, DefaultUserId, Req.tokenOrError(Req.AuthTokenKey))
-    authService.deleteProtectedResource(BaseUrl + "/" + id)
-    serialize(result)
+  override def destroy(id: String): EssentialAction = csrfCheckAction andThen authorisationAction async { implicit request =>
+    FlowApi.remove(id, DefaultUserId, Req.tokenOrError(Req.AuthTokenKey))
+      .map { response =>
+        if(response)
+          authService.deleteProtectedResource(BaseUrl + "/" + id)
+        response.toResult
+      }
   }
 
-  override def find(id: String): EssentialAction = csrfCheckAction { implicit request =>
-    serialize(FlowApi.instance(id, DefaultUserId, Req.tokenOrError(Req.AuthTokenKey)))
+  override def find(id: String): EssentialAction = csrfCheckAction async { implicit request =>
+    FlowApi.instance(id, DefaultUserId, Req.tokenOrError(Req.AuthTokenKey)).map(_.toResult)
   }
 
   override def create: EssentialAction = csrfCheckAction {
     NotImplemented
   }
 
-  def create(flowTemplateId: String): EssentialAction = (csrfCheckAction
-    andThen authorisationAction) { implicit request =>
-    val flowInstance: FlowInstance =
-      FlowApi.instantiate(flowTemplateId, DefaultUserId, Req.tokenOrError(Req.AuthTokenKey))
-    authService.createProtectedResource(
-      ScopeFlowInstanceView :: ScopeFlowInstanceUpdate :: ScopeFlowInstanceDelete :: Nil,
-      ResourceNamePrefix + ResourceNameSeparator + flowInstance.id,
-      BaseUrl + "/" + flowInstance.id,
-      Type,
-      authService.userId(request.claims)
-    )
-    serialize(flowInstance)
+  def create(flowTemplateId: String): EssentialAction = csrfCheckAction andThen authorisationAction async { implicit request =>
+    FlowApi.instantiate(flowTemplateId, DefaultUserId, Req.tokenOrError(Req.AuthTokenKey))
+      .map { flowInstance =>
+        authService.createProtectedResource(
+          ScopeFlowInstanceView :: ScopeFlowInstanceUpdate :: ScopeFlowInstanceDelete :: Nil,
+          ResourceNamePrefix + ResourceNameSeparator + flowInstance.id,
+          BaseUrl + "/" + flowInstance.id,
+          Type,
+          authService.userId(request.claims)
+        )
+        flowInstance.toResult
+      }
   }
 
-  def start(flowInstanceId: String): EssentialAction = csrfCheckAction { implicit request =>
-    serialize(FlowApi.start(flowInstanceId, DefaultUserId, Req.tokenOrError(Req.AuthTokenKey)))
+
+  def start(flowInstanceId: String): EssentialAction = csrfCheckAction async { implicit request =>
+    FlowApi.start(flowInstanceId, DefaultUserId, Req.tokenOrError(Req.AuthTokenKey)).map(_.toResult)
   }
 
-  def stop(flowInstanceId: String): EssentialAction = csrfCheckAction { implicit request =>
-    serialize(FlowApi.stop(flowInstanceId, DefaultUserId, Req.tokenOrError(Req.AuthTokenKey)))
+  def stop(flowInstanceId: String): EssentialAction = csrfCheckAction async { implicit request =>
+    FlowApi.stop(flowInstanceId, DefaultUserId, Req.tokenOrError(Req.AuthTokenKey)).map(_.toResult)
   }
 
   private def flowInstanceIds(permissions: List[Permission]): List[String] = {
@@ -85,7 +96,7 @@ class FlowInstanceApi @Inject()(csrfCheckAction: CSRFCheckAction,
 
     permissions
       .filter(p => p.resourceName.startsWith(ResourceNamePrefix + ResourceNameSeparator))
-        .map(p => p.resourceName.split(ResourceNameSeparator).tail.head)
+      .map(p => p.resourceName.split(ResourceNameSeparator).tail.head)
   }
 }
